+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 ### Changed
+- **BREAKING**: wrap `sdd::Guard` into `EbrGuard`.
 - Replace `scc` with `sdd` (EBR is moved to the dedicated crate).
 
 ## [0.1.1] - 2024-04-15